requirements.txt
# Core ML/DL
torch>=2.0.0
monai>=1.4.0
nibabel>=5.1.0
einops>=0.7.0
tqdm>=4.66.0
torchvision>=0.15.0

# Multi-Agent Framework
langgraph>=0.4.0
langchain-core>=0.3.0

# Image Processing
Pillow>=10.0.0

# LLM - Gemma-3
transformers==4.50.0
accelerate==1.4.0
sentencepiece==0.2.1

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Visualization & Imaging Metrics
matplotlib>=3.7.0
scikit-image>=0.21.0

# RAG / Knowledge Retrieval
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0

# Database (Patient History)
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0

# Logging
loguru>=0.7.0
python-dotenv>=1.0.0
